refactor: extract terminal context menu mixin
Refs #69
